Move OutcomeSpec to kernel and reduce visibility of BaseSpec
