+++ laws/shared/src/test/scala/cats/effect/BaseSpec.scala
@@ -26,7 +26,7 @@ import org.scalacheck.{Arbitrary, Gen, Prop}
 import org.scalacheck.util.Pretty
 
 // A dumping ground for random helpers for property tests
-trait BaseSpec {
+private[effect] trait BaseSpec {
 
   implicit def kleisliEq[F[_], A, B](implicit ev: Eq[A => F[B]]): Eq[Kleisli[F, A, B]] =
     Eq.by[Kleisli[F, A, B], A => F[B]](_.run)
